+++ src/modules/class/classroom.service.ts
@@ -1,4 +1,5 @@
 import {
+	BadRequestException,
 	ForbiddenException,
 	forwardRef,
 	Inject,
@@ -6,7 +7,11 @@ import {
 	InternalServerErrorException,
 	NotFoundException
 } from '@nestjs/common'
-import { compareNotesByRating } from '../notes/helper'
+import { SpaceType } from '../../common/constants'
+import { IConfigAttributes } from '../../common/interfaces/config/app-config.interface'
+import { getConfig } from '../../config'
+import { FilesService } from '../files/files.service'
+import { compareNotesWithCombinedFeatures } from '../notes/helper'
 import { Note } from '../notes/models/notes.model'
 import { ClassroomRepository } from './classroom.repository'
 import { ClassroomUser } from './models/classroom-user.model'
@@ -14,17 +19,54 @@ import { Classroom } from './models/classroom.model'
 import { User } from './models/user.model'
 import { UserService } from './user.service'
 
+const config: IConfigAttributes = getConfig()
+
 @Injectable()
 export class ClassroomService {
 	constructor(
 		private readonly classroomRepository: ClassroomRepository,
+		private readonly filesService: FilesService,
 		@Inject(forwardRef(() => UserService))
 		private readonly userService: UserService
 	) {}
 
-	async createClassroom(name: string, ownerId: number): Promise<Classroom> {
+	async deleteClassroomThumbnail(thumbUri: string): Promise<void> {
+		const thumbnailCustom = !(thumbUri === config.classThumbnailDefaultURI)
+		if (thumbnailCustom) {
+			// Delete the uploaded thumbnail
+			await this.filesService.deleteFileWithID(thumbUri, SpaceType.IMAGES)
+		}
+	}
+
+	async createClassroom(name: string, ownerId: number, thumbnailUri?: string): Promise<Classroom> {
 		// Create classroom
-		const cr: Classroom = await this.classroomRepository.insert(name, ownerId)
+		const thumbUri = thumbnailUri ? thumbnailUri : config.classThumbnailDefaultURI
+
+		// Ensure thumbnail exists on remote
+		const fileExists = await this.filesService.remoteFileExists(thumbUri, SpaceType.IMAGES)
+		if (!fileExists) {
+			throw new NotFoundException(`File with URI of ${thumbUri} does not exist. Try uploading again`)
+		}
+
+		// Confirm valid URI
+		const validImage = await this.filesService.isValidFileType(thumbUri, SpaceType.IMAGES)
+		if (!validImage) {
+			throw new BadRequestException(
+				`Invalid thumbnail format was submitted to create the classroom. Supported are (jpg, png)`
+			)
+		}
+
+		const cr: Classroom = await this.classroomRepository.insert(
+			name,
+			ownerId,
+			`https://${config.imageDataSpace}.${config.spacesEndpoint}/${thumbUri}`
+		)
+
+		// Ensure classroom object was created (inserted)
+		if (!cr) {
+			await this.deleteClassroomThumbnail(thumbUri)
+			throw new InternalServerErrorException(`Failed to create(insert) the classroom ...`)
+		}
 
 		// Join the owner to the classroom
 		const user: User = await this.userService.getUserWithID(ownerId)
@@ -35,6 +77,7 @@ export class ClassroomService {
 			const success: boolean = await this.deleteClassroom(cr.id, ownerId)
 			const crName: string = cr.name
 			if (!success) {
+				await this.deleteClassroomThumbnail(thumbUri) // Ensure thumbnail file deleted
 				throw new InternalServerErrorException(
 					`After failing to add ${user.firstName} to ${crName}, also failed to delete ${crName}`
 				)
@@ -99,7 +142,7 @@ export class ClassroomService {
 			throw new NotFoundException(`No notes were found in ${cr.name}`)
 		}
 
-		return notes.sort(compareNotesByRating)
+		return notes.sort(compareNotesWithCombinedFeatures)
 	}
 
 	async getClassroomUsers(classId: string, userId: number): Promise<User[]> {
@@ -137,7 +180,7 @@ export class ClassroomService {
 	async updateClassroom(
 		classId: string,
 		ownerId: number,
-		newData: { name?: string; ownerId?: number }
+		data: { name?: string; ownerId?: number; thumbnailUri?: string }
 	): Promise<Classroom> {
 		const cr: Classroom = await this.classroomRepository.get(classId)
 
@@ -149,7 +192,12 @@ export class ClassroomService {
 			throw new ForbiddenException(`You are not authorized to update this classroom (not owner)`)
 		}
 
-		return this.classroomRepository.update(cr, newData)
+		if (data.thumbnailUri) {
+			// Delete old thumbnail
+			await this.deleteClassroomThumbnail(cr.thumbnailUri)
+		}
+
+		return this.classroomRepository.update(cr, data)
 	}
 
 	async deleteClassroom(classId: string, ownerId: number): Promise<boolean> {
@@ -163,6 +211,8 @@ export class ClassroomService {
 			throw new ForbiddenException(`You are not authorized to delete this classroom (not owner)`)
 		}
 
+		// Ensure thumbnail is deleted with the classroom
+		await this.deleteClassroomThumbnail(cr.thumbnailUri)
 		return this.classroomRepository.delete(cr)
 	}