+++ src/modules/notes/helper/index.ts
@@ -1,28 +1,21 @@
-import { extname } from 'path'
-import { v4 as uuid } from 'uuid'
-import * as util from 'util'
-import * as fs from 'fs'
-import * as pdf from 'pdf-parse'
-import { IConfigAttributes } from 'src/common/interfaces/config/app-config.interface'
-import { getConfig } from '../../../config'
 import { Note } from '../models/notes.model'
+import { Rating } from 'src/modules/ratings/models/rating.model'
 
-const config: IConfigAttributes = getConfig()
-
-// Works only on PDF
-export async function extractBodyFromFile(path: string): Promise<string | undefined> {
-	const pathComponents: string[] = path.split('.')
-	if (pathComponents[pathComponents.length - 1] !== 'pdf') {
-		return 'BAD FORMAT (required: PDF)'
-	}
-
-	const readFile = util.promisify(fs.readFile)
-	const buffer = await readFile(`${config.fileStorageLocation}/${path}`)
-	const pdfData = await pdf(buffer, { max: 16 })
-
-	return pdfData.text
+/**
+ * Takes the full contents of a note upload and trims it to 50 words or less
+ * @param body The full contents of an uploaded note
+ * @returns A small abstract representative of the note
+ */
+export async function createNoteAbstract(body: string): Promise<string> {
+	const words = body.split(' ')
+	return `${words.slice(0, Math.min(50, words.length)).join(' ')} ...`
 }
 
+/**
+ * Calculates the average read time for a note
+ * @param body The body of the note being uploaded
+ * @returns A n average read time for the generic user
+ */
 export async function calculateReadTimeMinutes(body: string): Promise<number> {
 	// Source: https://irisreading.com/what-is-the-average-reading-speed/
 	const avgWordsPerMin = 200
@@ -30,46 +23,42 @@ export async function calculateReadTimeMinutes(body: string): Promise<number> {
 	return Math.round(body.split(' ').length / avgWordsPerMin) + 1
 }
 
-// Used to initialize a ratings format for a new note (stick to defaul)
-export function createEmptyRatings(ratingsSize = 5): number[] {
-	return Array(ratingsSize).fill(0)
+/**
+ * Get all ratings sorted by category for the purposes of comparing weighted ratings in *compareNotesWithCombinedFeatures()*
+ * @param ratings A list of valid rating objects
+ * @returns A compatible rating total by value for every rating in ratings
+ */
+export function getRatingTotals(ratings: Rating[]): number[] {
+	const totals: number[] = [ 0, 0, 0, 0, 0 ]
+	for (const r of ratings) {
+		totals[r.value - 1] += 1
+	}
+
+	return totals
 }
 
-export function editFileName(req, file: Express.Multer.File, cb) {
-	const name = file.originalname.split('.')[0]
-	const fileExtName = extname(file.originalname)
-	cb(null, `${name}-${uuid()}${fileExtName}`)
-}
+/**
+ * 
+ * @param a A Note object to be compared
+ * @param b A Note object to be compared
+ * @returns A number, positive Indicates that note B wins and negative indicates that note A wins
+ */
+export function compareNotesWithCombinedFeatures(a: Note, b: Note): number {
+	// Establish points (weights)
+	let aPoints = 0
+	let bPoints = 0
 
-// Used to compare notes by rating
-export function compareNotesByRating(a: Note, b: Note): number {
-	// Get the highest rating
-	const aRating = a.rating.indexOf(Math.max(...a.rating))
-	const bRating = b.rating.indexOf(Math.max(...b.rating))
+	// First find our which is newer
+	const aCreated = new Date(a.createdAt)
+	const bCreated = new Date(b.createdAt)
+	aCreated < bCreated ? (aPoints += 30) : (bPoints += 30)
 
-	// Get total ratings
-	const aTotal = a.rating.reduce((a, b) => a + b, 0)
-	const bTotal = a.rating.reduce((a, b) => a + b, 0)
+	// Now calculate points for each notes ratings
+	const aRatings: number[] = getRatingTotals(a.ratings)
+	const bRatings: number[] = getRatingTotals(b.ratings)
+	aPoints += aRatings.reduce((a, b, i) => a + b * i, 0)
+	bPoints += bRatings.reduce((a, b, i) => a + b * i, 0)
 
-	if (aRating < bRating) {
-		// Note b is higher rated
-		return 1
-	}
-	if (aRating > bRating) {
-		// Note a is higher rated
-		return -1
-	}
-
-	// equal rated (check counts)
-	if (aTotal < bTotal) {
-		// B has more ratings in total
-		return 1
-	}
-	if (aTotal > bTotal) {
-		// A has more ratings in total
-		return -1
-	}
-
-	// Complete tie
-	return 0
+	// Finally compare and return
+	return bPoints - aPoints
 }