+++ src/modules/notes/types/note-rating-response.type.ts
@@ -1,0 +1,6 @@
+import { ApiProperty } from '@nestjs/swagger'
+
+export class NoteRatingResponse {
+	@ApiProperty() statusCode: number
+	@ApiProperty() value: number
+}